feat: Add button and add chat model
