+++ .gitignore
@@ -1,2 +1,4 @@
 Meok-Kkae-Bi.xcodeproj/project.xcworkspace/xcuserdata
-
+.DS_Store
+Meok-Kkae-Bi.xcodeproj/xcworkspace
+Meok-Kkae-Bi.xcodeproj/xcuserdata
